+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - Added menu entries for pan cv input attenuation in tracks, groups and aux
 - Change "Mute track sends when group is muted" (in "Aux sends" menu, when AuxSpander is attached) to "Groups control track send levels" such that group fader, pan, mute, fade etc. control the track sends also
 - Implement fade in aux returns in AuxSpander
-
+- Add setting for buttons CV inputs that are of type Gate high/low, with original Toggle trigger as default setting
 
 ### 1.1.2 (2019-10-27)